modified controller to check login status

+++ app/Http/Controllers/TasksController.php
@@ -13,9 +13,17 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
+        
     {
+        $tasks =[];    
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
         //タスク一覧を取得
-        $tasks = Task::all();
+
+            $tasks = $user->tasks()->orderBy('id', 'asc')->get();
+        }        
+    
         // タスク一覧ビューで表示
         return view('tasks.index', [
             'tasks' => $tasks,
@@ -53,11 +61,13 @@ class TasksController extends Controller
         ]);
     
         // タスクを作成
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status =$request->status; //追加
-        $task->save();
-
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        
+        
         // トップページへリダイレクト
         return redirect('/');        //
     }
@@ -72,11 +82,17 @@ class TasksController extends Controller
     {
          // idでタスクを検索して取得
         $task = Task::findOrFail($id);
-
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を表示
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        return back();
         // タスク詳細でそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+//        return view('tasks.show', [
+//            'task' => $task,
+//        ]);
     }
 
     /**
@@ -89,11 +105,13 @@ class TasksController extends Controller
     {
         // idでタスクを検索して取得
         $task = Task::findOrFail($id);
-
-        // タスク編集Viewで表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            // タスク編集Viewで表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        return back();
     }
 
     /**
@@ -115,12 +133,16 @@ class TasksController extends Controller
         // idでタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを更新
-        $task->content = $request->content;
-        $task->status = $request->status; //追加
-        $task->save();
-
+        if (\Auth::id() === $task->user_id) {
+    
+            $task->content = $request->content;
+            $task->status = $request->status; //追加
+            $task->save();
         // トップページへリダイレクト
-        return redirect('/');
+
+            return redirect('/');
+        }
+        return back();    
     }
 
     /**
@@ -133,10 +155,12 @@ class TasksController extends Controller
     {
         // idでタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
-        // トップページへリダイレクト
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            // タスクを削除
+            $task->delete();
+            // トップページへリダイレクト
+            return redirect('/');
+        }
+        return back();
     }
 }
